Implement basic TLS support
Capable of securing both server-server and client-server communication. Uses the secret format used by cert-manager and kubectl, which consists of two fields: "tls.crt" and "tls.key". As mongod expects a single file with both of these we use an init container which reads the mounted secrets and combine them.