CLOUDP-74820: change in scram credential secret name generation (#234)
This PR changes the logic for SCRAM credential secret name generation for MongoDB users (to accommodate for non DNS1123 complying user name). Now a user is explicitly asked to provide a DNS1123 complying "secret name" which is being validated as opposed to the current operator behavior where it errors out during runtime when an invalid name is being generated for SCRAM secret(based on user name).